ExceptionToStatusInterceptor: Fix streaming responses
In streaming responses, ExceptionToStatusInterceptor noops, because the
intercept() method is called once and returns a generator before the
exception handling can take effect.

This change alters the intercept() method to additionally apply
exception handling inside the generator function it returns.

I'm not sure if this is the *correct* fix; it seems like maybe this
behaviour should live outside the intercept() function somehow. But it
fixes the issue for me and has a test.